+++ CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Timer(precision)` - Capture the time spent.
 -   `Duration(duration, precision)` - Represents a nanosecond duration.
 -   `iter_deep(*iterables)` - Iterator that returns elements from each iterable including nested ones.
+-   `default(value, default_value)` - Return the value if not None, or the default value.
+-   `format_heading(title, length, decorator, justify, margin, simple)` - Formats a heading.
+-   `format_columns(cells, length, separator, col_dir)` - Formats a list of cells into columns.
+-   `print_heading(title, length, decorator, justify, margin, simple)` - Prints a formatted heading.
+-   `print_columns(cells, length, separator, col_dir)` - Prints a list of cells into columns.
+-   `iter_cells(cells, nb_cols, nb_cells, col_dir)` - Iterator for returning elements from a list.
 
 ### Fixed